+++ lib/recording_studio_ai/models.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 require "recording_studio_ai/models/definition"
+require "recording_studio_ai/models/parameter_validation"
 require "recording_studio_ai/models/registry"